refact : rename class name and refact count player balls - use lambda
